// frontend/src/app/shared/generated-api/model/userSettingsUpdate.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSettingsUpdate { 
    theme?: UserSettingsUpdateThemeEnum;
    language?: UserSettingsUpdateLanguageEnum;
    moveCheckedToBottom?: boolean;
    trashAutoPurgeDays?: number;
}
export enum UserSettingsUpdateThemeEnum {
    Dark = 'dark',
    Light = 'light'
};
export enum UserSettingsUpdateLanguageEnum {
    En = 'en',
    De = 'de'
};